Implement enough for it to be (partially) working, with an initial basic renderer.
There seem to be some bugs keeping it from getting in-game, but at least
the attract sequence seems to be working.

For now, placeholder sprite rendering is in place, that just draws white
rectangles in place of the sprites. That's enough to see the code is
working, at least. The renderer is as barebones as possible, just to
check how well it's working.