Fix some bugs in the attract sequence graphics.
The attract sequence still has some graphical bugs, such as the
DEMONSTRATION/developer's website overlay strip not yet rendering
correctly. Also, there are still bugs in the rendering of the game,
which is used in the demo screens.

Some other minor improvements were made, such as changing the temporary
backdrop color to a light gray, so transparency can be debugged in the
future.